+++ src/controllers/usersController.js
@@ -136,13 +136,15 @@ const usersController = {
       // fs.writeFileSync(usersFilePath, JSON.stringify(users, null, ' '));
       // res.redirect('/users/login');
     } else {
-      fs.unlink(path.join(__dirname, "../../public/images/users/", req.file.filename),
+      if(req.file){
+        fs.unlink(path.join(__dirname, "../../public/images/users/", req.file.filename),
         (err) => {
           if (err) {
             console.log(err);
             return;
           }
         })
+      }
       res.render("./users/register", {
         errors: validations.errors,
         user: req.loggedUser,
@@ -251,13 +253,15 @@ const usersController = {
             });
         });
     } else {
-      fs.unlink(path.join(__dirname, "../../public/images/users/", req.file.filename),
+      if(req.file){
+        fs.unlink(path.join(__dirname, "../../public/images/users/", req.file.filename),
         (err) => {
           if (err) {
             console.log(err);
             return;
           }
         })
+      }
       res.render("./users/user", {
         errors: validations.errors,
         user: req.loggedUser,
